wasmer - extractOpcodeCost moved to opcodeCost.go
