+++ wasmer2/opcodeCost.go
@@ -6,135 +6,270 @@
 
 package wasmer2
 
+import "github.com/multiversx/mx-chain-vm-go/executor"
+
 type OpcodeCost struct {
-	Block                         uint32
-	Br                            uint32
-	BrIf                          uint32
-	BrTable                       uint32
-	Call                          uint32
-	CallIndirect                  uint32
-	Catch                         uint32
-	CatchAll                      uint32
-	Delegate                      uint32
-	Drop                          uint32
-	Else                          uint32
-	End                           uint32
-	GlobalGet                     uint32
-	GlobalSet                     uint32
-	I32Add                        uint32
-	I32And                        uint32
-	I32Clz                        uint32
-	I32Const                      uint32
-	I32Ctz                        uint32
-	I32DivS                       uint32
-	I32DivU                       uint32
-	I32Eq                         uint32
-	I32Eqz                        uint32
-	I32Extend16S                  uint32
-	I32Extend8S                   uint32
-	I32GeS                        uint32
-	I32GeU                        uint32
-	I32GtS                        uint32
-	I32GtU                        uint32
-	I32LeS                        uint32
-	I32LeU                        uint32
-	I32Load                       uint32
-	I32Load16S                    uint32
-	I32Load16U                    uint32
-	I32Load8S                     uint32
-	I32Load8U                     uint32
-	I32LtS                        uint32
-	I32LtU                        uint32
-	I32Mul                        uint32
-	I32Ne                         uint32
-	I32Or                         uint32
-	I32Popcnt                     uint32
-	I32RemS                       uint32
-	I32RemU                       uint32
-	I32Rotl                       uint32
-	I32Rotr                       uint32
-	I32Shl                        uint32
-	I32ShrS                       uint32
-	I32ShrU                       uint32
-	I32Store                      uint32
-	I32Store16                    uint32
-	I32Store8                     uint32
-	I32Sub                        uint32
-	I32WrapI64                    uint32
-	I32Xor                        uint32
-	I64Add                        uint32
-	I64And                        uint32
-	I64Clz                        uint32
-	I64Const                      uint32
-	I64Ctz                        uint32
-	I64DivS                       uint32
-	I64DivU                       uint32
-	I64Eq                         uint32
-	I64Eqz                        uint32
-	I64Extend16S                  uint32
-	I64Extend32S                  uint32
-	I64Extend8S                   uint32
-	I64ExtendI32S                 uint32
-	I64ExtendI32U                 uint32
-	I64GeS                        uint32
-	I64GeU                        uint32
-	I64GtS                        uint32
-	I64GtU                        uint32
-	I64LeS                        uint32
-	I64LeU                        uint32
-	I64Load                       uint32
-	I64Load16S                    uint32
-	I64Load16U                    uint32
-	I64Load32S                    uint32
-	I64Load32U                    uint32
-	I64Load8S                     uint32
-	I64Load8U                     uint32
-	I64LtS                        uint32
-	I64LtU                        uint32
-	I64Mul                        uint32
-	I64Ne                         uint32
-	I64Or                         uint32
-	I64Popcnt                     uint32
-	I64RemS                       uint32
-	I64RemU                       uint32
-	I64Rotl                       uint32
-	I64Rotr                       uint32
-	I64Shl                        uint32
-	I64ShrS                       uint32
-	I64ShrU                       uint32
-	I64Store                      uint32
-	I64Store16                    uint32
-	I64Store32                    uint32
-	I64Store8                     uint32
-	I64Sub                        uint32
-	I64Xor                        uint32
-	If                            uint32
-	LocalGet                      uint32
-	LocalSet                      uint32
-	LocalTee                      uint32
-	LocalAllocate                 uint32
-	Loop                          uint32
-	MemoryGrow                    uint32
-	MemorySize                    uint32
-	MemoryCopy                    uint32
-	MemoryFill                    uint32
-	Nop                           uint32
-	RefFunc                       uint32
-	RefIsNull                     uint32
-	RefNull                       uint32
-	Rethrow                       uint32
-	Return                        uint32
-	ReturnCall                    uint32
-	ReturnCallIndirect            uint32
-	Select                        uint32
-	TableGet                      uint32
-	TableGrow                     uint32
-	TableInit                     uint32
-	TableSet                      uint32
-	TableSize                     uint32
-	Throw                         uint32
-	Try                           uint32
-	TypedSelect                   uint32
-	Unreachable                   uint32
-	Unwind                        uint32
+	Block              uint32
+	Br                 uint32
+	BrIf               uint32
+	BrTable            uint32
+	Call               uint32
+	CallIndirect       uint32
+	Catch              uint32
+	CatchAll           uint32
+	Delegate           uint32
+	Drop               uint32
+	Else               uint32
+	End                uint32
+	GlobalGet          uint32
+	GlobalSet          uint32
+	I32Add             uint32
+	I32And             uint32
+	I32Clz             uint32
+	I32Const           uint32
+	I32Ctz             uint32
+	I32DivS            uint32
+	I32DivU            uint32
+	I32Eq              uint32
+	I32Eqz             uint32
+	I32Extend16S       uint32
+	I32Extend8S        uint32
+	I32GeS             uint32
+	I32GeU             uint32
+	I32GtS             uint32
+	I32GtU             uint32
+	I32LeS             uint32
+	I32LeU             uint32
+	I32Load            uint32
+	I32Load16S         uint32
+	I32Load16U         uint32
+	I32Load8S          uint32
+	I32Load8U          uint32
+	I32LtS             uint32
+	I32LtU             uint32
+	I32Mul             uint32
+	I32Ne              uint32
+	I32Or              uint32
+	I32Popcnt          uint32
+	I32RemS            uint32
+	I32RemU            uint32
+	I32Rotl            uint32
+	I32Rotr            uint32
+	I32Shl             uint32
+	I32ShrS            uint32
+	I32ShrU            uint32
+	I32Store           uint32
+	I32Store16         uint32
+	I32Store8          uint32
+	I32Sub             uint32
+	I32WrapI64         uint32
+	I32Xor             uint32
+	I64Add             uint32
+	I64And             uint32
+	I64Clz             uint32
+	I64Const           uint32
+	I64Ctz             uint32
+	I64DivS            uint32
+	I64DivU            uint32
+	I64Eq              uint32
+	I64Eqz             uint32
+	I64Extend16S       uint32
+	I64Extend32S       uint32
+	I64Extend8S        uint32
+	I64ExtendI32S      uint32
+	I64ExtendI32U      uint32
+	I64GeS             uint32
+	I64GeU             uint32
+	I64GtS             uint32
+	I64GtU             uint32
+	I64LeS             uint32
+	I64LeU             uint32
+	I64Load            uint32
+	I64Load16S         uint32
+	I64Load16U         uint32
+	I64Load32S         uint32
+	I64Load32U         uint32
+	I64Load8S          uint32
+	I64Load8U          uint32
+	I64LtS             uint32
+	I64LtU             uint32
+	I64Mul             uint32
+	I64Ne              uint32
+	I64Or              uint32
+	I64Popcnt          uint32
+	I64RemS            uint32
+	I64RemU            uint32
+	I64Rotl            uint32
+	I64Rotr            uint32
+	I64Shl             uint32
+	I64ShrS            uint32
+	I64ShrU            uint32
+	I64Store           uint32
+	I64Store16         uint32
+	I64Store32         uint32
+	I64Store8          uint32
+	I64Sub             uint32
+	I64Xor             uint32
+	If                 uint32
+	LocalGet           uint32
+	LocalSet           uint32
+	LocalTee           uint32
+	LocalAllocate      uint32
+	Loop               uint32
+	MemoryGrow         uint32
+	MemorySize         uint32
+	MemoryCopy         uint32
+	MemoryFill         uint32
+	Nop                uint32
+	RefFunc            uint32
+	RefIsNull          uint32
+	RefNull            uint32
+	Rethrow            uint32
+	Return             uint32
+	ReturnCall         uint32
+	ReturnCallIndirect uint32
+	Select             uint32
+	TableGet           uint32
+	TableGrow          uint32
+	TableInit          uint32
+	TableSet           uint32
+	TableSize          uint32
+	Throw              uint32
+	Try                uint32
+	TypedSelect        uint32
+	Unreachable        uint32
+	Unwind             uint32
+}
+
+func (wasmerExecutor *Wasmer2Executor) extractOpcodeCost(wasmOps *executor.WASMOpcodeCost) *OpcodeCost {
+	return &OpcodeCost{
+		Block:              wasmOps.Block,
+		Br:                 wasmOps.Br,
+		BrIf:               wasmOps.BrIf,
+		BrTable:            wasmOps.BrTable,
+		Call:               wasmOps.Call,
+		CallIndirect:       wasmOps.CallIndirect,
+		Catch:              wasmOps.Catch,
+		CatchAll:           wasmOps.CatchAll,
+		Delegate:           wasmOps.Delegate,
+		Drop:               wasmOps.Drop,
+		Else:               wasmOps.Else,
+		End:                wasmOps.End,
+		GlobalGet:          wasmOps.GlobalGet,
+		GlobalSet:          wasmOps.GlobalSet,
+		I32Add:             wasmOps.I32Add,
+		I32And:             wasmOps.I32And,
+		I32Clz:             wasmOps.I32Clz,
+		I32Const:           wasmOps.I32Const,
+		I32Ctz:             wasmOps.I32Ctz,
+		I32DivS:            wasmOps.I32DivS,
+		I32DivU:            wasmOps.I32DivU,
+		I32Eq:              wasmOps.I32Eq,
+		I32Eqz:             wasmOps.I32Eqz,
+		I32Extend16S:       wasmOps.I32Extend16S,
+		I32Extend8S:        wasmOps.I32Extend8S,
+		I32GeS:             wasmOps.I32GeS,
+		I32GeU:             wasmOps.I32GeU,
+		I32GtS:             wasmOps.I32GtS,
+		I32GtU:             wasmOps.I32GtU,
+		I32LeS:             wasmOps.I32LeS,
+		I32LeU:             wasmOps.I32LeU,
+		I32Load:            wasmOps.I32Load,
+		I32Load16S:         wasmOps.I32Load16S,
+		I32Load16U:         wasmOps.I32Load16U,
+		I32Load8S:          wasmOps.I32Load8S,
+		I32Load8U:          wasmOps.I32Load8U,
+		I32LtS:             wasmOps.I32LtS,
+		I32LtU:             wasmOps.I32LtU,
+		I32Mul:             wasmOps.I32Mul,
+		I32Ne:              wasmOps.I32Ne,
+		I32Or:              wasmOps.I32Or,
+		I32Popcnt:          wasmOps.I32Popcnt,
+		I32RemS:            wasmOps.I32RemS,
+		I32RemU:            wasmOps.I32RemU,
+		I32Rotl:            wasmOps.I32Rotl,
+		I32Rotr:            wasmOps.I32Rotr,
+		I32Shl:             wasmOps.I32Shl,
+		I32ShrS:            wasmOps.I32ShrS,
+		I32ShrU:            wasmOps.I32ShrU,
+		I32Store:           wasmOps.I32Store,
+		I32Store16:         wasmOps.I32Store16,
+		I32Store8:          wasmOps.I32Store8,
+		I32Sub:             wasmOps.I32Sub,
+		I32WrapI64:         wasmOps.I32WrapI64,
+		I32Xor:             wasmOps.I32Xor,
+		I64Add:             wasmOps.I64Add,
+		I64And:             wasmOps.I64And,
+		I64Clz:             wasmOps.I64Clz,
+		I64Const:           wasmOps.I64Const,
+		I64Ctz:             wasmOps.I64Ctz,
+		I64DivS:            wasmOps.I64DivS,
+		I64DivU:            wasmOps.I64DivU,
+		I64Eq:              wasmOps.I64Eq,
+		I64Eqz:             wasmOps.I64Eqz,
+		I64Extend16S:       wasmOps.I64Extend16S,
+		I64Extend32S:       wasmOps.I64Extend32S,
+		I64Extend8S:        wasmOps.I64Extend8S,
+		I64ExtendI32S:      wasmOps.I64ExtendI32S,
+		I64ExtendI32U:      wasmOps.I64ExtendI32U,
+		I64GeS:             wasmOps.I64GeS,
+		I64GeU:             wasmOps.I64GeU,
+		I64GtS:             wasmOps.I64GtS,
+		I64GtU:             wasmOps.I64GtU,
+		I64LeS:             wasmOps.I64LeS,
+		I64LeU:             wasmOps.I64LeU,
+		I64Load:            wasmOps.I64Load,
+		I64Load16S:         wasmOps.I64Load16S,
+		I64Load16U:         wasmOps.I64Load16U,
+		I64Load32S:         wasmOps.I64Load32S,
+		I64Load32U:         wasmOps.I64Load32U,
+		I64Load8S:          wasmOps.I64Load8S,
+		I64Load8U:          wasmOps.I64Load8U,
+		I64LtS:             wasmOps.I64LtS,
+		I64LtU:             wasmOps.I64LtU,
+		I64Mul:             wasmOps.I64Mul,
+		I64Ne:              wasmOps.I64Ne,
+		I64Or:              wasmOps.I64Or,
+		I64Popcnt:          wasmOps.I64Popcnt,
+		I64RemS:            wasmOps.I64RemS,
+		I64RemU:            wasmOps.I64RemU,
+		I64Rotl:            wasmOps.I64Rotl,
+		I64Rotr:            wasmOps.I64Rotr,
+		I64Shl:             wasmOps.I64Shl,
+		I64ShrS:            wasmOps.I64ShrS,
+		I64ShrU:            wasmOps.I64ShrU,
+		I64Store:           wasmOps.I64Store,
+		I64Store16:         wasmOps.I64Store16,
+		I64Store32:         wasmOps.I64Store32,
+		I64Store8:          wasmOps.I64Store8,
+		I64Sub:             wasmOps.I64Sub,
+		I64Xor:             wasmOps.I64Xor,
+		If:                 wasmOps.If,
+		LocalGet:           wasmOps.LocalGet,
+		LocalSet:           wasmOps.LocalSet,
+		LocalTee:           wasmOps.LocalTee,
+		LocalAllocate:      wasmOps.LocalAllocate,
+		Loop:               wasmOps.Loop,
+		MemoryGrow:         wasmOps.MemoryGrow,
+		MemorySize:         wasmOps.MemorySize,
+		Nop:                wasmOps.Nop,
+		RefFunc:            wasmOps.RefFunc,
+		RefIsNull:          wasmOps.RefIsNull,
+		RefNull:            wasmOps.RefNull,
+		Rethrow:            wasmOps.Rethrow,
+		Return:             wasmOps.Return,
+		ReturnCall:         wasmOps.ReturnCall,
+		ReturnCallIndirect: wasmOps.ReturnCallIndirect,
+		Select:             wasmOps.Select,
+		TableGet:           wasmOps.TableGet,
+		TableGrow:          wasmOps.TableGrow,
+		TableInit:          wasmOps.TableInit,
+		TableSet:           wasmOps.TableSet,
+		TableSize:          wasmOps.TableSize,
+		Throw:              wasmOps.Throw,
+		Try:                wasmOps.Try,
+		TypedSelect:        wasmOps.TypedSelect,
+		Unreachable:        wasmOps.Unreachable,
+		Unwind:             wasmOps.Unwind,
+	}
 }
